requirements.txt
streamlit
python-dotenv

beautifulsoup4
requests
pandas
schedule
selenium
